AB79431 refactored and added unit test
